requirements.txt
google-generativeai==0.8.5
PyGObject==3.50.0 
aiohttp==3.10.11
async-timeout==5.0.1
yarl==1.13.1
multidict==6.2.0
cachetools==5.3.3
google-api-core==2.19.0
protobuf==5.29.5
typing-extensions==4.15.0